requirements.txt
aiosmtplib==2.0.2
annotated-types==0.6.0
anyio
astroid
blinker==1.7.0
Brotli
certifi
cffi
charset-normalizer
click
colorama
coverage
cryptography~=42.0.2
deta==1.2.0
dill
dnspython==2.6.1
ecdsa
email_validator==2.1.1
fastapi~=0.103.0
fastapi-mail==1.4.1
future
h11
httpcore
httpx
idna
iniconfig
isort
Jinja2==3.1.3
lazy-object-proxy
mailersend==0.5.6
MarkupSafe==2.1.5
mccabe
mypy
mypy-extensions
packaging
platformdirs
pluggy
psutil
pyasn1
pycparser
pycryptodome
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
pylint
PySocks
pytest
python-dotenv==1.0.1
python-jose
python-multipart
requests
rsa
six
sniffio
starlette
tomlkit
types-pyasn1==0.5.0.20240205
types-python-jose==3.3.4.20240106
typing_extensions
urllib3
uvicorn
wrapt
pymysql

pymongo
sqlalchemy~=2.0.31